pyproject.toml
[tool.poetry]
name = "datacrunch-on-poetry"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
datacrunch = "^1.6.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
ipython = "^8.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
